add EnvKeys::validate() function
