+++ examples/using-slack-channel/app.php
@@ -8,7 +8,7 @@ declare(strict_types=1);
 require dirname(__DIR__) . '/../vendor/autoload.php';
 
 use Chemaclass\ScrumMaster\Channel\Slack;
-use Chemaclass\ScrumMaster\Common\Keys;
+use Chemaclass\ScrumMaster\Common\EnvKeys;
 use Chemaclass\ScrumMaster\IO\EchoOutput;
 use Chemaclass\ScrumMaster\IO\NotifierInput;
 use Chemaclass\ScrumMaster\IO\NotifierOutput;
@@ -19,14 +19,8 @@ use Symfony\Component\HttpClient\HttpClient;
 $dotEnv = Dotenv\Dotenv::create(__DIR__);
 $dotEnv->load();
 
-$mandatoryKeys = Keys::fromEnvFile(file_get_contents('./env.dist'));
-
-foreach ($mandatoryKeys as $mandatoryKey) {
-    if (!isset($_ENV[$mandatoryKey])) {
-        echo implode(', ', $mandatoryKeys) . 'keys are mandatory!';
-        exit(1);
-    }
-}
+$mandatoryKeys = EnvKeys::fromFile(file_get_contents(__DIR__ . '/.env.dist'));
+$mandatoryKeys->validate();
 
 $notifier = new Notifier(
     new JiraHttpClient(HttpClient::create([
